# public/docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Servidores",
        "description": "API Servidores",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/cidades": {
            "get": {
                "tags": [
                    "Cidades"
                ],
                "summary": "Lista todas as cidades",
                "operationId": "listCidades",
                "responses": {
                    "200": {
                        "description": "Lista de cidades"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Cidades"
                ],
                "summary": "Cria uma nova cidade",
                "operationId": "createCidade",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Cidade"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Cidade criada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Cidade"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Dados inválidos"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/cidades/{id}": {
            "get": {
                "tags": [
                    "Cidades"
                ],
                "summary": "Busca cidade por ID",
                "operationId": "getCidade",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cidade encontrada"
                    },
                    "404": {
                        "description": "Cidade não encontrada"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Cidades"
                ],
                "summary": "Atualiza uma cidade",
                "operationId": "updateCidade",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Cidade"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Cidade atualizada"
                    },
                    "404": {
                        "description": "Cidade não encontrada"
                    },
                    "400": {
                        "description": "Dados inválidos"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Cidades"
                ],
                "summary": "Exclui uma cidade",
                "operationId": "deleteCidade",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Cidade excluída"
                    },
                    "404": {
                        "description": "Cidade não encontrada"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/enderecos/{id}": {
            "get": {
                "tags": [
                    "Endereços"
                ],
                "summary": "Busca endereço por ID",
                "operationId": "getEndereco",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Endereço encontrado"
                    },
                    "404": {
                        "description": "Endereço não encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Endereços"
                ],
                "summary": "Atualiza endereço",
                "operationId": "updateEndereco",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Endereco"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Endereço atualizado"
                    },
                    "400": {
                        "description": "Dados inválidos"
                    },
                    "404": {
                        "description": "Endereço não encontrado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Endereços"
                ],
                "summary": "Exclui endereço",
                "operationId": "deleteEndereco",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Endereço excluído"
                    },
                    "404": {
                        "description": "Endereço não encontrado"
                    }
                }
            }
        },
        "/api/enderecos": {
            "post": {
                "tags": [
                    "Endereços"
                ],
                "summary": "Cria novo endereço",
                "operationId": "createEndereco",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Endereco"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Endereço criado"
                    },
                    "400": {
                        "description": "Dados inválidos"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Cidade": {
                "properties": {
                    "cid_id": {
                        "type": "integer",
                        "readOnly": true
                    },
                    "cid_nome": {
                        "type": "string"
                    },
                    "cid_uf": {
                        "type": "string",
                        "maxLength": 2
                    }
                },
                "type": "object"
            },
            "Endereco": {
                "properties": {
                    "end_id": {
                        "type": "integer",
                        "readOnly": true
                    },
                    "end_tipo_logradouro": {
                        "type": "string"
                    },
                    "end_logradouro": {
                        "type": "string"
                    },
                    "end_numero": {
                        "type": "integer"
                    },
                    "end_bairro": {
                        "type": "string"
                    },
                    "cid_id": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Laravel Sanctum token authentication",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
}